feat(crossing): refactoring moving out of tollgate folder
